[TEST] - GraphQLTest 에 인자가 필요한 요청을 테스트 하는 코드 추가
참고 :
https://medium.com/@philippechampion58/testing-your-graphql-apis-in-a-spring-boot-app-9fe02ebccc35

+++ src/main/java/com/sbkim/springboot_graphql/domain/Content.java
@@ -17,7 +17,7 @@ public class Content extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long contentId;
+    private int contentId;
 
     private String contentTitle;
